add age ranges to tests, re-create `LanguageLevel` model
a test can be defined for some age ranges; it can also be
designed for all levels of a language, which means that
language and levels must be selected separately.
That, in turn, means, that language level cannot just be
a `choices` field.
