+++ api/admin.py
@@ -14,6 +14,7 @@ for model in (
     models.Teacher,
     models.TeacherUnder18,
     models.Language,
+    models.LanguageLevel,
     models.LanguageAndLevel,
 ):
     admin.site.register(model)
